refactor: daily recommendation 타입 분리 #65
